refactor: don't match test cases with variable names
This changes the test case detection to omit cases where the name passed
to t.Run is a variable as currently parsing positions via treesitter
doesn't resolve these names to their final values (which is
a complicated ask) so it's not possible to run them properly
e.g. go tests -run '<variable>' does not work with using the actual
string value of variable.