feat: use --run flag when executing single test
Run the go test command with --run flag when executing single test.

Support for neotest `type=test` was removed in https://github.com/nvim-neotest/neotest-go/pull/72

This is unfortunate but i believe in wisdom of the maintainers. Anyway i
really need this functionality. Otherwise the plugin is unusable on
projects with e2e/integration tests. I can't execute the whole suite
each time, that is job for the CI.